+++ packages/react-native/src/input/index.tsx
@@ -88,7 +88,6 @@ const Input = forwardRef<TextInput, InputProps>(
         <Box flexGrow={1}>
           <TextInput
             ref={ref}
-            allowFontScaling={false}
             {...restProps}
             style={[
               {